+++ src/test/java/com/intellecteu/onesource/integration/api/EntityApiTestFactory.java
@@ -54,6 +54,7 @@ import com.intellecteu.onesource.integration.repository.entity.onesource.VenueEn
 import com.intellecteu.onesource.integration.repository.entity.onesource.VenuePartyEntity;
 import com.intellecteu.onesource.integration.repository.entity.toolkit.CloudSystemEventEntity;
 import com.intellecteu.onesource.integration.repository.entity.toolkit.DeclineInstructionEntity;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -115,7 +116,7 @@ public class EntityApiTestFactory {
             .venue(buildVenueEntity())
             .instrument(buildInstrumentEntity())
             .rate(buildRateEntity())
-            .quantity(2)
+            .quantity(BigDecimal.valueOf(2L))
             .billingCurrency(USD)
             .dividendRatePct(2d)
             .tradeDate(LocalDate.now())
